requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py.git@0bcb0d0e3ce395d42a5b1dae61b0090791ee018d
requests==2.28.0
websocket_client==1.3.3
